add readers, dashboard
