+++ app/reader/(dashboard)/layout.js
@@ -1,0 +1,10 @@
+import DashboardNav from "../(components)/DashboardNav";
+
+export default function Layout({ children }) {
+  return (
+    <main className="bg-gray-100">
+      <DashboardNav />
+      {children}
+    </main>
+  );
+}
